Add code to allow user to interact with game.
This update implements the portion of the game that allows a user to 
play the game. It allows users to click on cards to flip them over, 
detects matches, displays a winner animation at the end of the game, and 
then resets itself.

Broken:
The highlighting feature does not currently work properly. It is suppose 
to highlight a card when the mouse hovers over a card. It successfully 
hightlights the cards, but the highlight does not disappear when the 
mouse moves.